requirements.txt
numpy~=2.4.4
matplotlib~=3.10.8
qutip~=5.2.3
scipy~=1.17.1
zhinst-toolkit~=1.3.0